Eliminate compiler warnings

+++ logdefs.h
@@ -51,9 +51,9 @@
        buf[sizeof(buf)-1] = 0;
 #    ifndef __APPLE__
        if(!LogToSysLog) {
-	 fprintf(stderr,"[%ld] " LOG_MODULENAME "%s\n", syscall(__NR_gettid), buf);
+	 fprintf(stderr,"[%ld] " LOG_MODULENAME "%s\n", (long int)syscall(__NR_gettid), buf);
        } else {
-	 syslog(level, "[%ld] " LOG_MODULENAME "%s", syscall(__NR_gettid), buf);
+	 syslog(level, "[%ld] " LOG_MODULENAME "%s", (long int)syscall(__NR_gettid), buf);
        }
 #    else
        if(!LogToSysLog) {
